requirements.txt
# Core OCR Dependencies
easyocr>=1.7.0
opencv-python>=4.8.0
Pillow>=9.0.0

# API Framework
fastapi>=0.100.0
uvicorn>=0.20.0

# Data Validation
pydantic>=2.0.0

# Configuration Management
pyyaml>=6.0

# Logging and Utilities
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Development Dependencies
pytest>=7.0.0
pytest-asyncio>=0.20.0
pytest-benchmark>=4.0.0

# Image processing utilities
numpy>=1.21.0
scipy>=1.7.0

# Optional GPU Support (uncomment if needed)
# torch>=2.0.0
# torchvision>=0.15.0